Move Submitter class to backends.gcp module
